requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Pillow
google-genai
python-dotenv
pytest
